% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrome_colors.R
\name{anthrome_colors}
\alias{anthrome_colors}
\title{Anthrome color tables}
\usage{
anthrome_colors(kind = "class")
}
\arguments{
\item{kind}{The anthrome color palette to return. One of "class", "level",
or "type". Defaults to "class" the color palette for the full 21 anthrome
classes.}
}
\value{

}
\description{
Convenience functions for returning color tables for anthrome classes,
levels, and types
}
\examples{
anthrome_colors('class')
anthrome_colors('level')
anthrome_colors('type')
}
