% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{anthrome_legend_long}
\alias{anthrome_legend_long}
\title{Anthrome legends}
\usage{
anthrome_legend_long()
}
\value{

}
\description{
Convenience functions to load pre-rendered legends for anthrome plots to use
in lieu of those created by ggplot.
}
\examples{
anthrome_legend_long()
anthrome_legend_wide()

}
