% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrome_trajectory.R
\name{anthrome_summary}
\alias{anthrome_summary}
\title{Calculate anthrome trajectories}
\usage{
anthrome_summary(anthromes, inputs, by = NULL)
}
\arguments{
\item{by}{a column in data by which to group the anthrome data before summary}

\item{data}{a tibble containing the anthrome classification over time}
}
\value{

}
\description{
A convenience function to create summaries of the total land area in each
anthrome type at each time step, potentially subset by an additional
variable.
}
\examples{
# global summary
anthrome_trajectory(dat)
# summary by biome
anthrome_trajectory(dat, by = 'biome')

}
