% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrome_trajectory.R
\name{anthrome_trajectory}
\alias{anthrome_trajectory}
\title{Calculate Anthrome trajectories}
\usage{
anthrome_trajectory(data, by = NULL)
}
\arguments{
\item{data}{a tibble containing the anthrome classification over time}

\item{by}{a column in data by which to group the anthrome data before summary}
}
\value{

}
\description{
A convenience function to create summaries of the total land area in each
anthrome type at each time step, potentially subset by an additional
variable.
}
\examples{
# global summary
anthrome_trajectory(dat)
# summary by biome
anthrome_trajectory(dat, by = 'biome')
}
