% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{anthromes_med}
\alias{anthromes_med}
\title{Example datasets}
\format{
An object of class \code{stars} of dimension 121 x 121 x 6.
}
\usage{
anthromes_med
}
\description{
Anthromes classification for the Eastern Mediterranean, generated using
anthrome_classify() on hyde_med and inputs_med.
}
\examples{
data(inputs_med)
plot(inputs_med)
}
\keyword{datasets}
