% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgg_extract.R
\name{dgg_extract}
\alias{dgg_extract}
\title{Extract raster data with a DGG shapefile}
\usage{
dgg_extract(dat, dgg, var = NULL, fun)
}
\arguments{
\item{dat}{The raster data to extract.}

\item{dgg}{Shapefile containing the DGG boundaries of to use for the extraction}

\item{var}{The variable to extract.}
}
\value{

}
\description{
Extract raster data with a DGG shapefile
}
\examples{
dgg_extract(hyde_med, dgg_med, 'crops', 'sum')
}
