% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{dgg_med}
\alias{dgg_med}
\title{Example datasets}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7036 rows and 3 columns.
}
\usage{
dgg_med
}
\description{
Equal-area hexagonal discrete global grid system polygons, cropped to an area in
the Eastern Mediterranean.
}
\examples{
data(dgg_med)
plot(dgg_med)
}
\keyword{datasets}
