% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthromes_classify.R
\name{get_anthromes}
\alias{get_anthromes}
\title{Anthromes classification}
\usage{
get_anthromes(hyde, inputs)
}
\arguments{
\item{dat}{the tibble to apply the anthromes classification to. Must include
columns for total population count and areal fractions of urban, rice,
irrigation, crops, grazing lands as well as supporting data on potential
vegetation and potential village status.}
}
\value{
a tibble containing a new column for the resulting anthrome class
}
\description{
Applies the Anthromes-12k classification algorithm from Ellis et al. 2020 to a tibble of population
and land-use data.
}
\examples{
anthromes_classify(dat)

}
\references{
Ellis, E.C., Beusen, A.H. and Goldewijk, K.K., 2020.
Anthropogenic biomes: 10,000 BCE to 2015 CE. Land, 9(5), p.129.
}
