% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hyde.R
\name{get_hyde}
\alias{get_hyde}
\title{Get HYDE 3.2 population and land use data}
\usage{
get_hyde(
  vars = c("cropland", "grazing", "ir_rice", "popc", "tot_irri", "uopp"),
  dir = "."
)
}
\arguments{
\item{vars}{The HYDE 3.2 variables to import. Either any combination of
'cropland', 'grazing', 'ir_rice', 'popc', 'tot_irri', or 'uopp' for the associated
model output, or "inputs" for the HYDE 3.2 fixed inpus variables.}

\item{dir}{The directory where the compressed HYDE data are. If the file 'raw-data.zip'
is not currently in this directory, it will be downloaded from doi:10.7910/DVN/E3H3AK
on the Harvard Dataverse. Defaults to the current working directory.}
}
\value{
A 3-dimensional stars object in 5 arc minute geographic format format for all 75 time steps.
If vars = 'input', a 2-dimensional stars object.
}
\description{
Get HYDE 3.2 population and land use data
}
\examples{
get_hyde() # Imports all the time-varying HYDE 3.2 data and downloads if necessary.
get_hyde('inputs') # Imports the HYDE 3.2 fixed inputs
}
