% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyde_utils.R
\name{hyde2dgg}
\alias{hyde2dgg}
\title{Helper functions for dealing with HYDE data}
\usage{
hyde2dgg(hyde_varname, dgg, hyde_path)
}
\arguments{
\item{hyde_varname}{The HYDE 3.2 variable name to process.
One of 'cropland', 'grazing', 'ir_rice', 'popc', 'tot_irri', 'uopp'}

\item{dgg}{The discrete global grid file to use for extraction}

\item{hyde_path}{Path to HYDE 3.2 data}
}
\value{

}
\description{
A set of helper functions for dealing with HYDE data
}
\examples{
hyde2dgg('cropland', dgg_land, 'path-to-hyde')
}
