% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{hyde_med}
\alias{hyde_med}
\title{Example datasets}
\format{
An object of class \code{stars} of dimension 121 x 121 x 6.
}
\usage{
hyde_med
}
\description{
HYDE 3.2 land use and population data in stars format, cropped to an area in
the Eastern Mediterranean.
}
\examples{
data(hyde_med)
plot(hyde_med)
}
\keyword{datasets}
