% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{inputs_med}
\alias{inputs_med}
\title{Example datasets}
\format{
An object of class \code{stars} with 121 rows and 121 columns.
}
\usage{
inputs_med
}
\description{
HYDE 3.2 fixed input data in stars format, cropped to an area in
the Eastern Mediterranean.
}
\examples{
data(inputs_med)
plot(inputs_med)
}
\keyword{datasets}
