% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_data.R
\docType{data}
\name{anthrome_key}
\alias{anthrome_key}
\title{Supplemental data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 4 columns.
}
\usage{
anthrome_key
}
\value{

}
\description{
These are several convenience datasets that make it easier to plot and
analyze the anthromes data. They primarily include lookup tables to easily
translate between numeric and character representations of anthromes, biomes,
and other data, as well as color ramps for easy plotting.
}
\examples{
time_steps_centuries()
}
\keyword{datasets}
