% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrome_summary.R
\name{anthrome_summary}
\alias{anthrome_summary}
\title{Calculate anthrome summaries}
\usage{
anthrome_summary(anthromes, inputs, by = NULL)
}
\arguments{
\item{anthromes}{A stars object containing the anthrome classification.}

\item{inputs}{A stars object containing HYDE 3.2 fixed inputs.}

\item{by}{a column in data by which to group the anthrome data before summary}

\item{hyde}{A stars object containing HYDE 3.2 land use and population data.}
}
\value{

}
\description{
A convenience function to create summaries of the total land area in each
anthrome type at each time step, potentially subset by an additional
variable.
}
\examples{
# global summary
anthrome_summary(anthromes_med, inputs_med)
# summary by regions
anthrome_summary(anthromes_med, inputs_med, by = regions)

# summarise hyde data
hyde_summary(hyde_med, inputs_med, by = regions)
}
