% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgg_extract.R
\name{dgg_extract}
\alias{dgg_extract}
\title{Extract raster data with a DGG shapefile}
\usage{
dgg_extract(dat, dgg, var = NULL, fun)
}
\arguments{
\item{dat}{The raster data to extract.}

\item{dgg}{Shapefile containing the DGG boundaries to use for the extraction.}

\item{var}{The variable to extract.}

\item{fun}{The function to pass to exactextractr::exact_extract().}
}
\value{

}
\description{
Given a stars data object and shapefile, extracts the relevant data using
exactextractr::exact_extract(). Stars inputs should be in raster format with
x and y dimensions. This function is planned for deprecation whenever direct
support for exact_extract() is better supported via stars::aggregate().
}
\examples{
dgg_extract(hyde_med, dgg_med, 'crops', 'sum')
}
