% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthromes_classify.R
\name{get_anthromes}
\alias{get_anthromes}
\title{Anthromes classification}
\usage{
get_anthromes(hyde, inputs)
}
\arguments{
\item{hyde}{a stars object containing HYDE 3.2 data to apply the anthromes
classification to. Must include columns for total population count and
areal fractions of urban, rice, irrigation, crops, grazing lands as well as
supporting data on potential vegetation and potential village status.}

\item{inputs}{HYDE 3.2 fixed input data in stars format.}
}
\value{
a stars object containing the resulting anthrome classes with the
same dimensions as hyde.
}
\description{
Applies the Anthromes-12k classification algorithm from Ellis et al. 2020
to a stars object of population and land-use data.
}
\examples{
get_anthromes(hyde_med, inputs_med)

}
\references{
Ellis, E.C., Beusen, A.H. and Goldewijk, K.K., 2020.
Anthropogenic biomes: 10,000 BCE to 2015 CE. Land, 9(5), p.129.
}
